Fix mocked dates
